Refactoring: Move warmup state and adapter out of browser package.
