+++ ftw/monitor/configure.zcml
@@ -12,6 +12,10 @@
   <include package="zc.z3monitor" />
   <include package=".browser" />
 
+  <adapter
+      factory=".warmup.DefaultWarmupPerformer"
+      />
+
   <utility
       component=".health.health_check"
       provides="zc.z3monitor.interfaces.IZ3MonitorPlugin"
